backend/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
google-generativeai
pydantic-settings